closes #1802

<!-- You can skip this if you're fixing a typo or adding an app to the Showcase. -->

**Explain the *details* for making this change. What existing problem does the pull request solve?**

The PR solves the issue of parsing query params that contains array key in `[]` format. Following type of keys are parsed and the tests are added:

- `data[]=john&data[]=doe`
- `data[0][name]=john&data[0][age]=10&data[1][name]=doe&data[1][age]=12` // This will be parsed to a struct
- `data.0.name=john&data.0.age=10&data.1.name=doe&data.1.age=12`