@ninadingole 2022-03-16
I did the benchmarking:

**BEFORE:**
```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser -benchmem -count=4                                                              25.5m  Wed Mar 16 20:38:39 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser
Benchmark_Ctx_QueryParser-16          	  349766	      3171 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  364510	      3176 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  355460	      3161 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  361420	      3171 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  309963	      3514 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  309619	      3565 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  286914	      3519 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  328557	      3532 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	9.755s
```

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser_Comma -benchmem -count=4                                                            10.4s  Wed Mar 16 20:38:54 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  307441	      3551 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  335578	      3562 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  296473	      3554 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  327487	      3525 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	4.887s
```

**AFTER:**

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser -benchmem -count=4                                                                   5.5s  Wed Mar 16 20:39:39 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser
Benchmark_Ctx_QueryParser-16          	  335406	      3427 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  356060	      3241 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  342423	      3223 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  351063	      3250 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  316597	      3587 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  313954	      3590 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  308415	      3626 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  322690	      3647 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	9.642s
```

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser_Comma -benchmem -count=4                                                            10.2s  Wed Mar 16 20:40:05 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  297918	      3614 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  321032	      3583 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  323709	      3608 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  312877	      3606 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	4.915s
```


Benchmarking for the method `parseQuery`:

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_parseQuery -benchmem -count=4                                                                     5.5s  Wed Mar 16 20:44:41 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_parseQuery
Benchmark_Ctx_parseQuery-16    	  401638	      2637 ns/op	     753 B/op	      29 allocs/op
Benchmark_Ctx_parseQuery-16    	  451958	      2637 ns/op	     753 B/op	      29 allocs/op
Benchmark_Ctx_parseQuery-16    	  454078	      2664 ns/op	     753 B/op	      29 allocs/op
Benchmark_Ctx_parseQuery-16    	  432850	      2669 ns/op	     753 B/op	      29 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	5.246s
```