@ninadingole 2022-03-16
I also tried with `Replacer`

```
rep := strings.NewReplacer("[]", ".", "[", ".", "]", "")
```

**Benchmarking:**

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser -benchmem -count=4                                                                    6.4s  Wed Mar 16 20:44:53 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser
Benchmark_Ctx_QueryParser-16          	  335710	      3016 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  386149	      3042 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  363243	      3301 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser-16          	  377787	      3140 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  329350	      3717 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  337248	      3575 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  337785	      3409 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  335216	      3315 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	11.173s
```

```
go test -v  -run=^\$ -bench=Benchmark_Ctx_QueryParser_Comma -benchmem -count=4                                                            12.5s  Wed Mar 16 20:46:40 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16    	  316009	      3499 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  356570	      3519 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  292567	      3841 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_QueryParser_Comma-16    	  301339	      3866 ns/op	     929 B/op	      44 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	6.371s
```

**Benchmark the only method:**

```
 go test -v  -run=^\$ -bench=Benchmark_Ctx_parseQuery -benchmem -count=4                                                                    7.7s  Wed Mar 16 20:47:56 2022
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_Ctx_parseQuery
Benchmark_Ctx_parseQuery-16    	  211260	      5311 ns/op	    2371 B/op	      47 allocs/op
Benchmark_Ctx_parseQuery-16    	  232926	      5309 ns/op	    2371 B/op	      47 allocs/op
Benchmark_Ctx_parseQuery-16    	  212172	      5251 ns/op	    2371 B/op	      47 allocs/op
Benchmark_Ctx_parseQuery-16    	  210127	      5377 ns/op	    2371 B/op	      47 allocs/op
PASS
ok  	github.com/gofiber/fiber/v2	5.162s
```