@ninadingole 2022-03-16
> thx
> 
> pls check the comment from efectn
> 
> > Check linter and gosec error.

I handled the error handling, don't know if it is valid to `panic`, please let me know if the changes are good to merge the PR. Thank you!